feat: Image module including some AI features (#433)
* file abstraction

* image

* image endpoints

* image endpoints

* add file fetching

* file data in frontend

* update schema

* file module in frontend

* basic security for files

* wrapped file as image

* why is null never accepted :(

* fix not null problem

* start with rekognition

* label display

* cleanup

* add config and rename S3File back to File in comments and stringss

* new image module comment

* fix gps coordinate type

* fix linting issues

* small ux improvements that allow easier demo

* Apply suggestions from code review


* add message to i18n

* remove ownership teswt from service

* add old CognitoUser decorator

* satisfy linter

* Update backend/src/flox/modules/image/image.resolver.ts

+++ backend/src/flox/modules/file/file.resolver.ts
@@ -1,16 +1,49 @@
-import { Args, Resolver, Query, Mutation } from '@nestjs/graphql';
-import PublicFile from './entities/public_file.entity';
+import { Args, Mutation, Query, Resolver } from '@nestjs/graphql';
+import PublicFile from './entities/publicFile.entity';
 import { FileService } from './file.service';
 import { GetPublicFileArgs } from './dto/args/get-public-file.args';
 import { GetPrivateFileArgs } from './dto/args/get-private-file.args';
-import PrivateFile from './entities/private_file.entity';
+import PrivateFile from './entities/privateFile.entity';
 import { LoggedIn, Public } from '../auth/authentication.decorator';
-import { User } from '../auth/entities/user.entity';
 import { DeleteFileInput } from './dto/input/delete-file.input';
+import { GetAllFilesArgs } from './dto/args/get-all-files.args';
+import { AdminOnly, CurrentUser } from '../roles/authorization.decorator';
+import { User } from '../auth/entities/user.entity';
+import { DEFAULT_ROLES } from '../roles/config';
+import { ForbiddenError } from 'apollo-server-express';
 
 @Resolver(() => PublicFile)
 export class FileResolver {
   constructor(private readonly fileService: FileService) {}
+
+  /**
+   * Returns all public files stored within database
+   * @param {GetAllFilesArgs} getAllFilesArgs - contains limit and skip parameters
+   * @returns {Promise<PublicFile[]>} List of public files
+   */
+  @AdminOnly()
+  @Query(() => [PublicFile], { name: 'allPublicFiles' })
+  async getAllPublicFiles(
+    @Args() getAllFilesArgs: GetAllFilesArgs,
+  ): Promise<PublicFile[]> {
+    return this.fileService.getAllPublicFiles(getAllFilesArgs);
+  }
+
+  /**
+   * Returns private files of logged in user
+   * @param {GetAllFilesArgs} getAllFilesArgs - contains limit and skip parameters
+   * @param {User} user - currently logged in user
+   * @returns {Promise<PrivateFile[]>} Users private files
+   */
+  @LoggedIn() // TODO application specific: set appropriate guards here
+  @Query(() => [PrivateFile], { name: 'allMyFiles' })
+  async getAllMyFiles(
+    @Args() getAllFilesArgs: GetAllFilesArgs,
+    @CurrentUser() user: User,
+  ): Promise<PrivateFile[]> {
+    console.log(user);
+    return this.fileService.getAllMyFiles(getAllFilesArgs, user);
+  }
 
   /**
    * Gets a public file
@@ -18,7 +51,7 @@ export class FileResolver {
    * @returns {Promise<PublicFile>} - the file
    */
   @Public()
-  @Query(() => PublicFile, { name: 'getPublicFile' })
+  @Query(() => PublicFile, { name: 'publicFile' })
   async getPublicFile(
     @Args() getPublicFileArgs: GetPublicFileArgs,
   ): Promise<PublicFile> {
@@ -28,28 +61,42 @@ export class FileResolver {
   /**
    * Gets a private file
    * @param {GetPrivateFileArgs} getPrivateFileArgs - contains UUID and optionally, expiration time
+   * @param {User} user - logged-in user
    * @returns {Promise<PrivateFile>} - the file, if the user is allowed to access it
    */
   @LoggedIn() // TODO application specific: set appropriate guards here
-  @Query(() => PrivateFile, { name: 'getPrivateFile' })
+  @Query(() => PrivateFile, { name: 'privateFile' })
   async getPrivateFile(
     @Args() getPrivateFileArgs: GetPrivateFileArgs,
+    @CurrentUser() user: User,
   ): Promise<PrivateFile> {
-    return this.fileService.getPrivateFile(getPrivateFileArgs);
+    const file = await this.fileService.getPrivateFile(getPrivateFileArgs);
+    if (user.role !== DEFAULT_ROLES.ADMIN || file.owner !== user.uuid) {
+      throw new ForbiddenError('File does not belong to logged in user');
+    }
+    return file;
   }
 
   /**
    * Deletes a private file
    * @param {DeleteFileInput} deleteFileInput - contains UUID
+   * @param {User} user - logged-in user
    * @returns {Promise<PrivateFile>} - the file that was deleted
    */
   @LoggedIn() // TODO application specific: set appropriate guards here
-  @Mutation(() => User)
+  @Mutation(() => PrivateFile)
   async deletePrivateFile(
-    @Args('deleteFileInput')
-    deleteFileInput: DeleteFileInput,
+    @Args('deleteFileInput') deleteFileInput: DeleteFileInput,
+    @CurrentUser() user: User,
   ): Promise<PrivateFile> {
-    // TODO application specific: Ensure only allowed person (usually admin or file owner) is allowed to delete
+    const file = await this.fileService.getPrivateFile({
+      uuid: deleteFileInput.uuid,
+    } as GetPrivateFileArgs);
+    if (user.role !== DEFAULT_ROLES.ADMIN || file.owner !== user.uuid) {
+      throw new ForbiddenError(
+        'Cannot delete file that does not belong to user',
+      );
+    }
     return this.fileService.deleteFile(
       deleteFileInput,
       false,
@@ -61,13 +108,12 @@ export class FileResolver {
    * @param {DeleteFileInput} deleteFileInput - contains UUID
    * @returns {Promise<PrivateFile>} - the file that was deleted
    */
-  @LoggedIn() // TODO application specific: set appropriate guards here
-  @Mutation(() => User)
+  @AdminOnly() // TODO application specific: set appropriate guards here
+  @Mutation(() => PublicFile)
   async deletePublicFile(
     @Args('deleteFileInput')
     deleteFileInput: DeleteFileInput,
   ): Promise<PublicFile> {
-    // TODO application specific: Ensure only allowed person (usually admin ) is allowed to delete
     return this.fileService.deleteFile(
       deleteFileInput,
       true,
